requirements.txt
# Web Framework
Flask==3.1.3
Werkzeug==3.1.6

# Database Connector สำหรับ MariaDB
mysql-connector-python==9.6.0
# หรือถ้าทีมถนัดใช้ ORM สามารถใช้ Flask-SQLAlchemy แทนได้

# AI Model & Data Processing
tensorflow==2.16.1      
numpy==1.26.4
Pillow==12.1.1          
opencv-python==4.9.0.80 # สำหรับฟีเจอร์ประมวลผลภาพเดิม (ถ้ายังต้องใช้)
scikit-learn==1.8.0
python-dotenv==1.2.1    # สำหรับจัดการไฟล์ .env (เก็บรหัสผ่าน Database)